# specs/api/openapi/ivr-order-confirmation.v1.yaml
openapi: 3.1.0
info:
  title: IVR Order Confirmation — Internal/Admin API (Target V1 Draft)
  version: 1.0.0-draft.4
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >
    Target V1 draft for the standalone .NET IVR service. Not a public API. The
    Sales Platform owns eligibility, revalidation, and order-state transitions.
    Supports GOLDEN_HOUR+ONLINE and TWENTY_FOUR_SEVEN+COD. External Sales callback
    is specified separately in order-core-ivr-callback.target-v1.yaml.
servers:
  - url: /v1/ivr/order-confirmation
    description: Internal base path
security:
  - bearerAuth: []
...
paths:
...
  /call-jobs:
    post:
      tags: [internal]
      summary: Create/record CallJob
      operationId: createCallJob
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/InternalSourceSystem'
        - $ref: '#/components/parameters/InternalServiceScope'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CallJobLifecycleRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrCallJob' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    get:
      tags: [admin]
      summary: Call-job list, masked and paginated (perm IVR_QUEUE_VIEW)
      description: >-
        `order_code` is a filter input only and is never echoed; the response
        carries the script-approved `order_code_short` (D-05, specs/ui/02).
      operationId: listCallJobs
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
        - { name: program, in: query, required: false, schema: { $ref: '#/components/schemas/ProgramCode' } }
        - { name: status, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: queue_status, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: result_type, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: order_code, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: correlation_id, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: near_expiry, in: query, required: false, schema: { type: boolean, default: false } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrCallJobPage' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
  /queue:
    get:
      tags: [admin]
      summary: Queue/capacity projection (masked)
      operationId: getQueue
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrQueueProjection' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /dashboard:
    get:
      tags: [admin]
      summary: Operational dashboard projection, masked (perm IVR_QUEUE_VIEW)
      description: >-
        Aggregates are computed server-side so the console never derives a KPI
        from raw rows. SIM pool and open incidents are pool-wide and ignore the
        program/time filter.
      operationId: getDashboard
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
        - { name: program, in: query, required: false, schema: { $ref: '#/components/schemas/ProgramCode' } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrDashboardProjection' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /call-jobs/{ivrCallJobId}/detail:
    get:
      tags: [admin]
      summary: Full masked lifecycle trace for one call job (perm IVR_QUEUE_VIEW)
      description: >-
        task intake to eligibility to attempts to result to callback, plus the
        technical-exception and review-item ids the admin mutations require.
        `order_state` is the opaque Core enum; IVR displays it and never changes
        it (D-02).
      operationId: getCallJobDetail
      security: [ { bearerAuth: [] } ]
      parameters:
        - { name: ivrCallJobId, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrCallJobDetail' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /scripts:
    get:
      tags: [admin]
      summary: Call-script catalogue, read-only (perm IVR_QUEUE_VIEW)
      description: >-
        Versions with their approval matrix, the allowed and prohibited variable
        lists, and the DTMF map. KEY_9 is reported NOT_ENABLED (AS-07). Lifecycle
        transitions are not exposed: approval is an owner action under OD-V1-15,
        not a console button.
      operationId: getScriptCatalog
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrScriptCatalog' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /integration-status:
    get:
      tags: [admin]
      summary: Downstream dependency status, read-only (perm IVR_QUEUE_VIEW)
      description: >-
        Reports what IVR can observe about itself and marks every unprobed
        dependency NOT_WIRED. `dependency_probing_available` stays false until
        P6-1 (W-0040) delivers real probing, so no card may be read as verified
        fail-closed before then.
      operationId: getIntegrationStatus
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
        - { name: environment, in: query, required: false, schema: { type: string, maxLength: 40 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrIntegrationStatus' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /review-items:
    get:
      tags: [admin]
      summary: Human review queue, masked and paginated (perm IVR_QUEUE_VIEW)
      operationId: listReviewItems
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ActorId'
        - { name: status, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrReviewQueue' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /queue:pause:
    post:
      tags: [admin]
      summary: Pause queue (perm IVR_QUEUE_PAUSE)
      operationId: pauseQueue
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminMutationRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrAdminActionResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /queue:resume:
    post:
      tags: [admin]
      summary: Resume queue (perm IVR_QUEUE_RESUME)
      operationId: resumeQueue
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminMutationRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrAdminActionResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /sim-channels/{simChannelId}:disable:
    post:
      tags: [admin]
      summary: Disable SIM channel (perm IVR_SIM_DISABLE)
      operationId: disableSim
      parameters:
        - name: simChannelId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminMutationRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrAdminActionResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /sim-channels/{simChannelId}:enable:
    post:
      tags: [admin]
      summary: Enable SIM channel after health pass (perm IVR_SIM_ENABLE)
      operationId: enableSim
      parameters:
        - name: simChannelId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminMutationRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrAdminActionResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /technical-retries:
    post:
      tags: [admin]
      summary: Request technical retry (perm IVR_MANUAL_RETRY; not a customer attempt)
      operationId: technicalRetry
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TechnicalRetryRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrTechnicalRetryResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /admin-reviews:
    post:
      tags: [admin]
      summary: Record review/annotation (perm IVR_RESULT_REVIEW)
      operationId: adminReview
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminReviewRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IvrAdminReviewResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Service token (allowlist Order Core, DF-06) or admin session (DF-01).
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
    ActorId:
      name: X-Actor-Id
      in: header
      required: true
      schema: { type: string }
...
    InternalSourceSystem:
      name: X-Source-System
      in: header
      required: true
      schema: { type: string, enum: [ivr-worker, ivr-adapter] }
    InternalServiceScope:
      name: X-Service-Scope
      in: header
      required: true
      schema: { type: string, enum: [ivr.internal.write] }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ProgramCode:
      type: string
      enum: [GOLDEN_HOUR, TWENTY_FOUR_SEVEN]
...
    CallJobLifecycleRequest:
      type: object
      additionalProperties: false
      required: [ivr_call_job_id, task_id]
      properties:
        ivr_call_job_id: { type: string, minLength: 1 }
        task_id: { type: string, minLength: 1 }
...
    IvrCallJob:
      type: object
      additionalProperties: false
      required:
        - ivr_call_job_id
        - task_id
        - program_type
        - status
        - queue_status
        - eligible
        - max_attempts
        - expires_at
        - input_signal_only
        - no_direct_order_update
      properties:
        ivr_call_job_id: { type: string }
        task_id: { type: string }
        official_order_id: { type: string }
        order_version: { type: string, description: Required Target V1 snapshot. }
        program_type: { $ref: '#/components/schemas/ProgramCode' }
        status: { type: string }
        attempt_policy_version: { type: string }
        max_attempts: { type: integer, minimum: 1, maximum: 10 }
        attempt_spacing_seconds: { type: integer }
        confirmation_window_seconds: { type: integer, minimum: 1 }
        queue_status: { type: string }
        eligible: { type: boolean }
        expires_at: { type: string, format: date-time }
        input_signal_only: { type: boolean, enum: [true] }
        no_direct_order_update: { type: boolean, enum: [true] }
...
    IvrQueueProjection:
      type: object
      additionalProperties: false
      required:
        - paused
        - pending_jobs
        - active_attempts
        - enabled_channels
        - open_hold_incidents
        - projected_at
      properties:
        paused: { type: boolean }
        pending_jobs: { type: integer, minimum: 0 }
        active_attempts: { type: integer, minimum: 0 }
        enabled_channels: { type: integer, minimum: 0 }
        open_hold_incidents: { type: integer, minimum: 0 }
        projected_at: { type: string, format: date-time }
...
    IvrDashboardProjection:
      type: object
      additionalProperties: false
      required:
        - generated_at
        - execution_mode
        - sim_provider
        - real_customer_call_allowed
        - queue
        - results
        - attempts
        - sim
        - open_incidents
        - missed_deadline_count
      properties:
        generated_at: { type: string, format: date-time }
        execution_mode: { type: string }
        sim_provider: { type: string }
        real_customer_call_allowed: { type: boolean }
        program_filter: { $ref: '#/components/schemas/ProgramCode' }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        queue: { $ref: '#/components/schemas/IvrDashboardQueuePanel' }
        results: { $ref: '#/components/schemas/IvrDashboardResultPanel' }
        attempts: { $ref: '#/components/schemas/IvrDashboardAttemptPanel' }
        sim: { $ref: '#/components/schemas/IvrDashboardSimPanel' }
        open_incidents:
          type: array
          items: { $ref: '#/components/schemas/IvrCapacityIncidentSummary' }
        missed_deadline_count: { type: integer, minimum: 0 }
...
    IvrCallJobPage:
      type: object
      additionalProperties: false
      required: [page, page_size, total_count, items]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total_count: { type: integer, minimum: 0 }
        items:
          type: array
          items: { $ref: '#/components/schemas/IvrCallJobListItem' }
...
    IvrCallJobDetail:
      type: object
      additionalProperties: false
      required:
        - ivr_call_job_id
        - task_id
        - order_code_short
        - phone_masked
        - program_type
        - order_state
        - order_version_snapshot
        - status
        - queue_status
        - eligible
        - eligibility_decision
        - blocked_reasons
        - call_restriction
        - max_attempts
        - attempt_policy_code
        - script_version
        - privacy_policy_version
        - t0_at
        - expires_at
        - created_at
        - attempts
        - results
        - callbacks
        - technical_exceptions
        - review_items
        - evidence_refs
        - audit_refs
        - correlation_id
        - input_signal_only
        - no_direct_order_update
      properties:
        ivr_call_job_id: { type: string }
        task_id: { type: string }
        order_code_short: { type: string, maxLength: 40 }
        phone_masked: { type: string }
        program_type: { $ref: '#/components/schemas/ProgramCode' }
        order_state:
          type: string
          description: Opaque enum owned by Order Core (D-02); displayed, never derived.
        order_version_snapshot: { type: string }
        status: { type: string }
        queue_status: { type: string }
        eligible: { type: boolean }
        eligibility_decision: { type: string }
        blocked_reasons:
          type: array
          items: { type: string }
        call_restriction: { type: boolean }
        sellable_captured_at: { type: string, format: date-time }
        max_attempts: { type: integer, minimum: 0 }
        attempt_policy_code: { type: string }
        script_version: { type: string }
        privacy_policy_version: { type: string }
        t0_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time }
        closed_reason: { type: string }
        attempts:
          type: array
          items: { $ref: '#/components/schemas/IvrCallAttemptDetail' }
        results:
          type: array
          items: { $ref: '#/components/schemas/IvrCallResultDetail' }
        callbacks:
          type: array
          items: { $ref: '#/components/schemas/IvrResultCallbackDetail' }
        technical_exceptions:
          type: array
          items: { $ref: '#/components/schemas/IvrTechnicalExceptionDetail' }
        review_items:
          type: array
          items: { $ref: '#/components/schemas/IvrReviewItemDetail' }
        evidence_refs:
          type: array
          items: { type: string }
        audit_refs:
          type: array
          items: { type: string }
        correlation_id: { type: string }
        input_signal_only: { type: boolean, enum: [true] }
        no_direct_order_update: { type: boolean, enum: [true] }
...
    IvrScriptCatalog:
      type: object
      additionalProperties: false
      required:
        - generated_at
        - execution_mode
        - production_target_v1_fields_approved
        - allowed_input_fields
        - prohibited_variables
        - dtmf_map
        - required_approval_types
        - versions
      properties:
        generated_at: { type: string, format: date-time }
        execution_mode: { type: string }
        production_target_v1_fields_approved:
          type: boolean
          description: OD-V1-15. While false the Target V1 field set is not production approved.
        allowed_input_fields:
          type: array
          items: { type: string }
        prohibited_variables:
          type: array
          items: { type: string }
        dtmf_map:
          type: array
          items: { $ref: '#/components/schemas/IvrDtmfKey' }
        required_approval_types:
          type: array
          items: { type: string }
        versions:
          type: array
          items: { $ref: '#/components/schemas/IvrScriptVersion' }
...
    IvrReviewQueue:
      type: object
      additionalProperties: false
      required: [page, page_size, total_count, items]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total_count: { type: integer, minimum: 0 }
        items:
          type: array
          items: { $ref: '#/components/schemas/IvrReviewQueueItem' }
...